updated the way models are marked as modified and added in things to handle the timezone
