+++ src/ukrdc_cupid/core/investigate/create_investigation.py
@@ -1,3 +1,8 @@
+
+
+
+
+
 def create_invalid_feed(patients:tuple):
     print("placeholder function for investigation into matched patients where demographics aren't validated to local feed")
     return "my_sqla_model_1"
@@ -8,6 +13,7 @@ def create_ambiguous_feed(patients:tuple):
 
 
 def create_invalid_ukrdc(pid:str):
+    # The
     print("placeholder function for investigation into ukrdc matched patients where demographics aren't validated")
     return "my_sqla_model_3"
